+++ .~lock.Untitled 1.odt#
@@ -1,0 +1,1 @@
+,brock,szu-email,25.01.2021 08:46,file:///home/brock/.config/libreoffice/4;
